+++ src/database/referrers.js
@@ -5,6 +5,7 @@ const dateWithOffset = require('../utils/dateWithOffset')
 
 const {
 	REFERRERS_SORTING_TOP,
+	REFERRERS_SORTING_NEW,
 	REFERRERS_SORTING_RECENT
 } = require('../constants/referrers')
 
@@ -33,6 +34,47 @@ const getTop = async (id) => {
 		{
 			$sort: {
 				count: -1
+			}
+		},
+		{
+			$limit: 25
+		}
+	])
+
+}
+
+const getNew = async (id) => {
+
+	return Record.aggregate([
+		{
+			$match: {
+				domainId: id,
+				siteReferrer: {
+					$ne: null
+				}
+			}
+		},
+		{
+			$group: {
+				_id: '$siteReferrer',
+				count: {
+					$sum: 1
+				},
+				created: {
+					$first: '$created'
+				}
+			}
+		},
+		{
+			$match: {
+				created: {
+					$gte: dateWithOffset(-6)
+				}
+			}
+		},
+		{
+			$sort: {
+				created: -1
 			}
 		},
 		{
@@ -78,6 +120,7 @@ const get = async (id, sorting) => {
 
 	switch (sorting) {
 		case REFERRERS_SORTING_TOP: return getTop(id)
+		case REFERRERS_SORTING_NEW: return getNew(id)
 		case REFERRERS_SORTING_RECENT: return getRecent(id)
 	}