+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] - 2019-11-05
+
+### Added
+
+- "New referrers"
+- Custom tracker URL (#53)
+
+### Fixed
+
+- Incorrect content type for JS files (#54)
+
 ## [1.3.0] - 2019-10-19
 
 ### Added